requirements.txt
PyMuPDF
pytesseract
Pillow
spacy
sentence_transformers
faiss-cpu
streamlit
gdown